helpers exec cleanup
- dropped unused functions
- moved helpers file used only in tests
