+++ acceptance/helpers/exec.go
@@ -1,0 +1,33 @@
+// Copyright © 2021 - 2023 SUSE LLC
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//     http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package helpers
+
+import (
+	"os"
+)
+
+// CreateTmpFile creates a temporary file on the disk with the given contents
+// and returns the path to it and an error if something goes wrong.
+func CreateTmpFile(contents string) (string, error) {
+	tmpfile, err := os.CreateTemp("", "epinio")
+	if err != nil {
+		return tmpfile.Name(), err
+	}
+	if _, err := tmpfile.Write([]byte(contents)); err != nil {
+		return tmpfile.Name(), err
+	}
+	if err := tmpfile.Close(); err != nil {
+		return tmpfile.Name(), err
+	}
+
+	return tmpfile.Name(), nil
+}
